% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-trafo-helpers.R
\name{as_ped.pam_xgb}
\alias{as_ped.pam_xgb}
\alias{as_ped.ped}
\title{Transform newdata to PED format based on fitted xgb model}
\usage{
\method{as_ped}{pam_xgb}(data, newdata, ...)

\method{as_ped}{ped}(data, newdata, ...)
}
\description{
Transform newdata to PED format based on fitted xgb model
}
