% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-trafo-helpers.R
\name{as_xgb_newdata}
\alias{as_xgb_newdata}
\title{Transform new data to xgb-PED format using same arguments as previous transformation}
\usage{
as_xgb_newdata(x, newdata, ...)
}
\description{
Transform new data to xgb-PED format using same arguments as previous transformation
}
\keyword{internal}
