% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-helpers.R
\name{get_cindex}
\alias{get_cindex}
\title{C-Index from pec package}
\usage{
get_cindex(
  object,
  data,
  cind_params = list(formula = Surv(time, status) ~ 1, exact = FALSE, start = 0.01),
  eval_times = NULL,
  q_last = 0.8,
  q_eval = c(0.25, 0.5, 0.75)
)
}
\description{
C-Index from pec package
}
