% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-splits.R
\name{get_cv_indices}
\alias{get_cv_indices}
\title{Create a lst of train/test indices}
\usage{
get_cv_indices(y, nfold = 4)
}
\arguments{
\item{nfold}{number of folds for cross-validation.}

\item{n}{number of rows in data set.}
}
\description{
Create a lst of train/test indices
}
\keyword{internal}
