% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-helpers.R
\name{get_ibs}
\alias{get_ibs}
\title{Extract IBS in tidy format}
\usage{
get_ibs(
  object,
  data,
  pec_params = list(formula = Surv(time, status) ~ 1, exact = FALSE, start = 0.01),
  keep_reference = TRUE,
  q_last = 0.8,
  q_eval = c(0.25, 0.5, 0.75),
  ...
)
}
\arguments{
\item{object}{A named list of prediction models, where allowed entries are
(1) R-objects for which a \link[pec]{predictSurvProb} method exists (see
details), (2) a \code{call} that evaluates to such an R-object (see
examples), (3) a matrix with predicted probabilities having as many rows as
\code{data} and as many columns as \code{times}. For cross-validation all
objects in this list must include their \code{call}.}

\item{data}{A data frame in which to validate the prediction models and to
fit the censoring model.  If \code{data} is missing, try to extract a data
set from the first element in object.}

\item{keep_reference}{Logical. Should Brier Score of Reference model be kept
in results table.}

\item{...}{Not used.}
}
\description{
Extract IBS in tidy format
}
