% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-splits.R
\name{get_split_indices}
\alias{get_split_indices}
\title{Split data into training and eval data for watchlist}
\usage{
get_split_indices(n, split_frac = 0.7)
}
\description{
Split data into training and eval data for watchlist
}
\keyword{internal}
