% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictSurvProb.pam_xgb}
\alias{predictSurvProb.pam_xgb}
\title{S3 method for compatibility with package pec}
\usage{
\method{predictSurvProb}{pam_xgb}(object, newdata, times)
}
\description{
S3 method for compatibility with package pec
}
