% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-splits.R
\name{random_params}
\alias{random_params}
\title{Draw random parameters from prespecified parameter space}
\usage{
random_params(param_space)
}
\description{
Checks if integer or double. Samples integers or draws from uniform distribution
respectively.
}
