% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-helpers.R
\name{times_list}
\alias{times_list}
\title{Look up Quantiles of event times in data}
\usage{
times_list(
  data,
  q_last = 0.8,
  q_eval = c(0.25, 0.5, 0.75),
  time_var = "time",
  status_var = "status",
  status_value = 1
)
}
\description{
Look up Quantiles of event times in data
}
\keyword{internal}
