% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost-fit.R
\name{xgb.tune_pam}
\alias{xgb.tune_pam}
\title{Tune xgb pam}
\usage{
xgb.tune_pam(
  data,
  param_df,
  nrounds,
  cv_indices,
  nfold = 4,
  ped_params = list(),
  nthread = 1L,
  early_stopping_rounds = NULL,
  verbose = FALSE,
  print_every_n = 1L,
  ...
)
}
\arguments{
\item{data}{training dataset. \code{xgb.train} accepts only an \code{xgb.DMatrix} as the input.
\code{xgboost}, in addition, also accepts \code{matrix}, \code{dgCMatrix}, or name of a local data file.}

\item{param_df}{A data frame of parameter combinations to tune. One
row contains one parameter set that will be passed on to
\code{params} in \code{xgb.cv_pam}.}

\item{nrounds}{max number of boosting iterations.}

\item{nfold}{Number of cross-valdation folds.}

\item{ped_params}{List of parameters used to transform data into PED format.}

\item{early_stopping_rounds}{If \code{NULL}, the early stopping function is not triggered.
If set to an integer \code{k}, training with a validation set will stop if the performance
doesn't improve for \code{k} rounds.
Setting this parameter engages the \code{\link[xgboost]{cb.early.stop}} callback.}

\item{verbose}{If 0, xgboost will stay silent. If 1, it will print information about performance.
If 2, some additional information will be printed out.
Note that setting \code{verbose > 0} automatically engages the
\code{cb.print.evaluation(period=1)} callback function.}

\item{print_every_n}{Print each n-th iteration evaluation messages when \code{verbose>0}.
Default is 1 which means all messages are printed. This parameter is passed to the
\code{\link[xgboost]{cb.print.evaluation}} callback.}

\item{...}{other parameters to pass to \code{params}.}
}
\description{
Tune xgb pam
}
