% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost-fit.R
\name{xgboost.ped}
\alias{xgboost.ped}
\title{Fit xgboost model to piece-wise exponential data}
\usage{
xgboost.ped(data = NULL, ...)
}
\arguments{
\item{data}{An objet of class \code{ped}.}

\item{...}{Further arguments passed to \code{\link[xgboost]{xgboost}}.}
}
\description{
Transforms a data set of class \code{ped} to a format suitable for
\code{xgboost}, then runs the xgboost model. All parameters available in
\code{\link[xgboost]{xgboost}} can be passed on via ellipsis.
The \code{base_score} is set to \code{1}.
}
