% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-trafo-helpers.R
\name{as_xgb_data}
\alias{as_xgb_data}
\alias{as_xgb_data.data.frame}
\alias{as_xgb_data.ped}
\title{Transform data suitable for xgboost applications}
\usage{
as_xgb_data(x, ...)

\method{as_xgb_data}{data.frame}(x, features, label, set_margin = TRUE, ...)

\method{as_xgb_data}{ped}(x, ...)
}
\description{
Transform data suitable for xgboost applications
}
